Horse Gear Fixes
Reverted Armor Types back to regular and added a Caparison type. Buffs for each Horse type reduced to only one and weighted with rpg_buff_weight. Vanilla was poorly scaled, mine are the same values as they were, but with less excessive editing and pointless tiered buffs when it could be accomplished easier through the current system of rpg_buff_weight.