More Fixes
Forgot to add Horse Caparison to be able to be repaired and that I had weapons set to breakable, which can make unique one time items and quest items break. May redo that if I take the time to flag only the worthy ones, though I personally would rather that be an all or none.